Weapon improvements / fixes
Nuke ops can now buy the AK-922 and its magazine. Nuke ops can now buy
the Glock 17 and it's magazine. Fixed some issues with the Abzats' ammo
boxes and etc for nuke ops. Vastly decreased the cost of the Abzats for
nuke ops. Various other improvements and fixes for weapons.