This updates the pool with new stuff.
OBJECTS CAN BE REMOVED IF PULLED.
It adds wires to it.
It makes people in the pool imune to rad storm.
It heals LOW ammount of damage if used properly (If you have more damage than the treshold, it dosen't heal you)
Treshold 30 tox, 20 fire, 10 brute. Proper use is tied to staminaloss while in the pool - You gotta keep moving, but not too much. It's delicate.