Quick fix for issue #189
Did some research and for some reason the the protected_jobs list is not
getting transfered over to the restricted_jobs list at round start as it
should. So i just put the protected jobs list into the restricted jobs
list for the time being... Would like to find out why the setting is not
being correctly transferred eventually.

Research steps included: Setting limits to 1 (as my server would only
have 1 person)
Forcing round to always traitor.
Enrolling as cyborg only (The only role on the restricted_jobs list)
Then restarting the round 10 times.
Test concluded that I was not set to antag a single time, therefore the
restricted_jobs list was working properly, and the protected_jobs list
was not.