Big pool update.
The pool no longer EMPs anything, feel free to swim with guns.
The poolcontroller has now a better UI, and allows for it to be drained/filled if wires are hacked.
The Filter now fills up by item stolen by the drain. You can empty the filter with a crowbar.
The Drain is dangerous when active, and can gib very damaged bodies.
You can now have water splashes fights.
 A lot fixed with the jumping board.