New bee related items
A new spell for the wizard, Dr. BEES costume for the costume vending
machine, two new giant bee mobs, new drink that involves Honey and Gin.
Also fixes Conjuration sounds, adds in a new randomization system that
can be applied to certain spells, and the ability to add sounds to
previously silent spells.