G17 and Abzats polishing update.
This update adds a new set of sprites for the Abzats to make it more
distinguishable from the L6 SAW. The G17 is now compatible with
suppressors and has an improved sprite for the magazine. Also (possibly)
fixes some guns being able to have suppressors attached when it wasn't
intended. The G17 is also properly named in the traitor / nuke ops
uplink.