Science chem lab map update
Updates the new chemistry lab in science to be more secure and harder to
break into. Also organizes the storage room in the bottom left of the
area to make it more accessible. Additionally, some minor adjustments to
the magazines of the AK-922 and the AK-922 icons themselves to reflect
on the casings now being dark-green.