Saber / Prototype SMG Magazine Nerf
Nerfs the SABER's / Prototype SMG's (made in RnD) magazines so that they
can only hold 30 rounds of 9mm instead of 42. In Pre-rebase this was
something like 21 rounds max but it was really meh in practice.
Furthermore, the fire_delay of the SMG was changed to 1, as this was the
value in Pre-Rebase code. Also makes the magazines with exotic ammo
(such as armor-piercing, incendiary, etc) come in a lockbox. Regular
ammo can still be printed E-Z P-Z. Alongside that, some of the magazines
no longer require glass in their production because that made no sense
at all. Finally fixes a ton of typos in the protolathe descriptions of
the magazines.