Unloading or loading weapon magazines now plays sounds
Every gun that uses magazines have been given their very own sound
effects which play when the magazine is loaded into or unloaded out of
the gun! In addition, chambering (removing) a round from the gun with no
magazine in also plays its own sound.
Shotguns which load via a tube now play a SFX when a shell is loaded.
The total size of the sound folder which adds all these new SFX is 199
KB. Credits to Insurgency (standalone) and Receiver for the sound
effects.