New Mac 10, .44 Automag, C05-R, and P057A Luger
Added 4 new guns! These guns are the Mac-10 SMG chambered in 9mm, C05-R
(Mauser C96) chambered in .45 ACP, P057A Luger (PO-8 Luger) chambered in
.357, and .44 Automag chambered in .44 AMP. Credits to Nienhaus for the
majority of the sprites.
All guns except for the automag are currently summon-guns exclusive. A
single automag spawns on the bar counter with the nuke ops leader. It
has a 7-round magazine and does a very formidable amount of damage. Also
has some armor-piercing traits which makes it dangerous even to users of
a bulletproof vest. The ammo that comes with the automag is the only
ammo that nuke ops can ever get for it, so it is advised that you
conserve your ammo.