New cat grenade, fedora tipping action button
You can now use an action button to tip a fedora if you're wearing this.
Doing this too much is not recommended by your local Chief Medical
Officer. Detective's fedora is also a subtype of the regular fedora now.
Added a grenade for traitors that spawns 8 feral cats when detonated.
Costs 5 TC.