Ion Inc. PMC Updates (more gear)
New equipment added for the Ion faction, specifically the PMC units.
This new equipment is a set of urban fatigues, heavy full-body combat
armor, PMC SWAT Helmet, and XMG80 assault rifle.

The armor is among some of the strongest in the game, ranking close or
equal to that of deathsquad armor. The XMG80 assault rifle fires
6.8x43mm caseless ammunition from a 48-round magazine. The sprite is
based on a G36 but in a bullpup configuration for those curious. Thanks
to Sgt. Fragg for the XMG80 sprites, and thanks to Fury McFurry for the
armor and clothing sprites.

Note for admins, you can easily give a full set of this gear to any mob
through the select equipment option and by equipping them either with
the Ion Inc. PMC Rifleman or Field Technician outfit.

A picture of all the new gear:
http://puu.sh/lOPR6/a84abdad76.png
Note: The fully-clothed one on the left is using the rifleman loadout,
while the one on the right is using the field technician loadout.

The new gunshot SFX for the XMG80 specifically (sounds even better in
game):
http://puu.sh/lOQwN/531a8189f2.ogg