Brig Update V3
Finished the security chutes around the station to the holding cell,
redesigned the Det Office, added in a darkroom. Few minor tweaks to fix
some issues in Brig such as Firelocks overlaying doors and Cell 3 not
working. Also put some more materials in Engineering at the request of
Derp to more realistically reflect the fact they're relied on to repair
the station.