{ORIGINAL AUTHOR IS KILM} Arrivals Overhaul
Complete Arrivals overhaul. This area now has a customs feature whereby
Sec can, if they want, man it and search those joining the station late.
It just adds a bit more depth to a historically underused area of the
station and gives potential for antag use.
The old Arrivals sec checkpoint is now the Invasive Search Area
Sorted out no power to the two APC's in the Brig