Nuke Ops Overhaul Part 1: Blastco
Mainly an addition of the (Very fucking rough.) base for the rest of the
nuke ops teams.

This part adds Blast Co, each of them get a new hardsuit (Sprites
courtesy of Wjohn)
A grenade launcher which slows them down and some ammo , a macrobomb, a
drop beacon for a normal bomb,  2 minibombs, 3 c4's, an incen grenade,
rollerskates to speed you up, and a pinpointer.

Also adds a sniper and chem bundle for nuke ops, and fixes slowdown for
inhands.