Moonlight Greatsword for Wizards
Added the Moonlight Greatsword for wizards. To use it, it must be
charged from damaging non-crit humanoids 5 times. Once charged, you can
use it as a ranged weapon to fire 5 pellets dealing 15 brute each. After
firing, it must be fully charged once again. It can fit on your back.
All credits to FuryMcFury for the sprites.

http://puu.sh/mpPDy/897cf17e4b.png

http://puu.sh/mpPEu/91dbbdd4f0.png

http://puu.sh/mpPEU/3394b58195.png

http://puu.sh/mpPF7/090cdfe17c.png