Stun abusability fix
For very small doses, Crank, Bath Salts and Meth will not stun and will only deal a bit of stamina damage. This is to prevent it from being abused as an instant stun and knockout in syringe guns.

Also reduced Meth's stamina damage from 50 to 25, as 50 would render you almost immobile after the stun.