Replaced "random" type cable coils
Now you can use the parent type and if you don't specify a colour on
creation it will pick one at random.

I've updated the maps to take care of this change. I tested that the
cables spawn in chemistry (which is two random colours) and the YouTool
in engineering also spawning random coloured wires.

I've also tested the wires so you can add them to the stack if you're
cyborg and the wires are not ones you placed down.