Changed to 'relaymove' procedure
It has a new optional parameter "delay". If this isn't specified it'll
check if the item has a definition; If it does it uses that (e.g. 10 for
office chair, 4 for wheelchairs, etc.), otherwise it uses default value
of 10.

Also added variables to office chair and wheelchair for delay and
reverse_direction flag