Completely revamped the upgrade
Made some major changes to how the the securitron boost can be enabled:

* Created a "Bot Upgrade: Boost" circuit board which requires some
research in order to create.
* Added an upgrade list so all bots (i.e. medibot, cleanbot, etc.) can
potentially make use of the upgrade (and more!)
* Added a method of added the circuit board to bots; Screwdriver open
the maintenance panel, add the card in
* Refactored how the boost was handled within securitron so it's now
more OOP than previous. Further work may be done

This is still a WIP as I need to confirm R&D works with this and do some
more checking when you're trying to upgrade a bot