Initial commit
Refactored how a lot of the screaming works. The screaming sound now can
be overridden with an "alternate_screams" list present in all /carbon/
types.

If this list is empty then the scream is the default one (e.g. male for
human male, cyborg for cyborg, etc.). However if this list has any items
in it then a random sound is picked when screaming. Thank you Dan Hibiki
for this suggestion!

I've added some procs to /carbon/ and /human/ which will allow you to
add and remove screams using a list. This will open more the possibility
of having other items that alter your scream. What this means is
equipping special items that have an alternate scream will add to the
list of alternate screams you can use.

I've also added the Cosby sweater to the premium section of the Clothes
Mate vending machine.