Creates moderator rank
Moderator is a lower rank than admin. Has access to basic tools such as
admin chat, and the ability to ban. Unlike trial admin, cannot spawn
things, heal dead people, or var edit things, so they are much less able
to screw things up.

Intended to be a way for an admin to come on and promote somebody
trustworthy (a mentor probably) if they have to go and they're the last
one.