fixed, also a list of stuff this does
This turns hulks green using alien color and grayscale sprites. To do
this, it adds grayscale moth parts as well, that's why it changes
human_parts_grayscale or whatever the file is.

Adds an icon for x ray vision that is a sort of blinking aura around the
eyes.

Makes the cold resistance aura look a bit better when limbs are missing
but this could still be improved.

Fixes eye lasers from the wizard's mutate power being sideways.

Adds in an angry animu style throbbing vein to separate hulk aliens from
plain old green ones.