Adds admin logs anytime an admin ghosts or views the player panel
Admins can ghost and see a shit ton of round information and it's not
even logged very easy to get away with. In the same vein, an Admin can
open the player panel, hover over someone's name and see if they are a
traitor or not, absolutely no way to get caught.  This fixes it, so if
an admin admin ghosts or opens the player panel, or shows the player
panel of a specific player, other admins are notified.  That's all.