Ups the prize of Disintegrate and Plasma Fist
Is there anything to add? Disintegrate requires barely any time to hit someone with and is a instant win with a stun. Plasma Fist, while technically is only better disintegrate since no one uses the other ones, still requires the boost. I'd change the mechanics of Plasma Fist too, like only being able to gib if the target is in critical health, but I can't do that changing a single file.