Makes gearing up as sec/command related jobs less of a pain
So this PR does a lot of minor things that make gearing up as Sec a lot
easier:

- Security belts now start out loaded with handcuffs, stun baton,
flashbang, pepperspray, and sec light.

- All security now start with a flash and handcuffs in their left and
right pockets. They lose the baton.

- Removes flashes, stun batons, pepper spray, and sec light from sec
lockers.

- Removes telescopic baton from the HoS backpack places it in is his
locker if he wants to use it (most don't) and removes the stun baton
from his locker (it's in the belt)

- Adds three black gloves to the Security wardrobe

- Changes the order of the items in lockers to make the first item in
the locker different for clicking and equipping as fast as possible:

- Sec lockers are now (HUDS, security belt, and helmet)

- Captain is now (HUD's, gloves, energy gun)

- HoS is now (HoS gun, belt, telescopic items, loyalty implants)

-HoP is now (sunglasses, gun, flash)

I think that's about it.