Vampire Overhaul (WIP) and Shadowlings
So the purpose of this overhaul may be a bit controversial, I'm not sure
if it'll ever get finished because of that.

Basically, it seems like to me that every experienced player knows that
shadowling is a broken game mode. Shadowlings to me seem like a fan
fiction tier version of vampires. Vampires are kind of bare bones right
now, and kind of lame. So uh.. the eventual goal of this PR is to get
rid of Shadowling entirely and replace it with the vampire game mode.
Some select shadowling powers will be given to the vampires.

Vampires ideally will work like they do on Goon and /vg/ with a mix of
how they work on Hippie.

- Start off with a few powers
- Drain clean blood to gain more powers
- Try to make as many vampires as possible.

So uh I'm thinking of giving the vampires something like: Lesser Glare
and nerfed shadow walk at the start.. haven't thought much out beyond
that.
So uh I'm thinking of giving the vampires something like: :