The Makarov Update
Adds the Makarov PM and PB.

The Standard Makarov hold 8 rounds of 9mm. Standard magazines are loaded
with AP, along with toxin and incindiary as special ammo.

The Makarov PB is the same as the standard makarov, but with an integral
suppressor. It can also be named with a pen.

Added a syndicate bundle containing the standard Makarov, three standard
magazines, a combat knife, soviet uniform, combat boots, conbat gloves,
ushanka, agent ID, voice changer mask, and a bottle of vodka. Minus the
clothes, the pack is worth 19 TC. With the clothes, probably closer to
25.

Added a mime only uplink kit containing a Makarov PM, two toxin mags,
and one incindiary mag for 14 TC. 10 for the gun, 2 for the toxin mags,
2 for the incindiary mag.