Ninja Rework - Stage 2
Compressed even more shit, making sure to leave little love notes in
places where the compression might break things. Said notes contain the
filename that the code used to belong in. Once this is finished, I'll
just be unticking the old ninja code just in case, so the notes are just
to guide future coders towards the ninja module counterpart for
reference.