Ninja Rework: Part 3
As it is, the code is fully functional on at least the same level as it
was before, maybe even better. It's still a work in progress, haven't
ironed out the server shitting itself with multiple ninjas, but it's a
significantly cleaner mess. The code can be merged now as a bug fix
while I work on the actual overhaul, or this PR can stay up while I
finish the rest.