Taser Stun Duration Test
Take 2: Electric Boogaloo, now on the live repo.

Reduces duration of stuns from the electrode projectile from 5 to 3. Should reduce the stun times enough that a beaten down or tired target will be down just long enough to cuff, but anyone else will need either a melee stun or  krav maga pin. This is just a short one or two week test to see how the playerbase reacts to to changes in the combat meta, so that we can collect data for large scale overhauls. This test will show us how the playerbase changes the meta to adapt to the changes, as well as how they react beyond combat. The end goal is an overhaul to ranged stuns which will please the majority of players, without completely alienating the powergamers and their cults of personality to the point that the server population drops.